commit Independent Processors
